Add viewer-based credential support for Azure, Databricks & Snowflake.
This commit brings support for Posit Connect's "viewer-based
credentials" feature [0] to Azure, Databricks, and Snowflake chatbots.

Checks for viewer-based credentials are designed to fall back gracefully
to existing authentication methods. This is intended to allow users to
-- for example -- develop and test a Shiny app that uses Azure,
Databricks, or Snowflake credentials in desktop Positron/RStudio or
Posit Workbench and deploy it with no code changes to Connect.

Most of the actual work is outsourced to a new shared package,
`connectcreds` [1]. In this latest version, the API of that package has
been dramatically simplified, meaning the changes to `ellmer` are far
less invasive than in previous patches.

Unit tests are included. They make use of the mocking features of the
`connectcreds` package to emulate a Connect environment.

[0]: https://docs.posit.co/connect/user/oauth-integrations/
[1]: https://github.com/posit-dev/connectcreds/

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>