Detect credentials from Azure service principals in chat_azure().
Inspired by a recent comment [0], this commit adds support for picking
up on Azure service principals, which seem to be a common way to furnish
`chat_azure()` with credentials without the need for an API key.

This approach also handles caching and expiry for these tokens, an
ongoing source of problems for users.

Unit tests are included. I've also updated the documentation.

[0]: https://github.com/tidyverse/ellmer/issues/262#issuecomment-2606628246

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>